requirements.txt
fastapi==0.95.2
uvicorn==0.18.3
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.2
python-multipart==0.0.5
jinja2==3.1.2